Move VXZoom into its own file for easier testing. Start tests
