--- packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/withVXZoom.spec.js
@@ -1,2 +1,0 @@
-import * as d3 from 'd3'
-import React from 'react'
